Refactor chart categories (partial)

+++ flowcentral-configuration/src/main/java/com/flowcentraltech/flowcentral/configuration/constants/ChartCategoryDataType.java
@@ -39,7 +39,7 @@ public enum ChartCategoryDataType implements EnumConst {
     LONG(
             "LGN",
             "numeric",
-            Long.class),
+            Integer.class),
     DATE(
             "DTE",
             "datetime",
